add wezterm

+++ gui/pkg.nix
@@ -6,17 +6,20 @@
   ...
 } @ args: let
   nixGLWrap = pkg:
-    pkgs.runCommand "${pkg.name}-nixgl-wrapper" {} ''
-      mkdir $out
-      ln -s ${pkg}/* $out
-      rm $out/bin
-      mkdir $out/bin
-      for bin in ${pkg}/bin/*; do
-       wrapped_bin=$out/bin/$(basename $bin)
-       echo "exec ${lib.getExe pkgs.nixgl.nixGLIntel} $bin \$@" > $wrapped_bin
-       chmod +x $wrapped_bin
-      done
-    '';
+    if withGLHack
+    then
+      pkgs.runCommand "${pkg.name}-nixgl-wrapper" {} ''
+        mkdir $out
+        ln -s ${pkg}/* $out
+        rm $out/bin
+        mkdir $out/bin
+        for bin in ${pkg}/bin/*; do
+         wrapped_bin=$out/bin/$(basename $bin)
+         echo "exec ${lib.getExe pkgs.nixgl.nixGLIntel} $bin \$@" > $wrapped_bin
+         chmod +x $wrapped_bin
+        done
+      ''
+    else pkg;
 in {
   fonts.fontconfig.enable = withGui;
 
@@ -26,6 +29,7 @@ in {
 
   imports = [
     (import ./kitty.nix (args // {inherit nixGLWrap;}))
+    (import ./wezterm.nix (args // {inherit nixGLWrap;}))
     ./element.nix
     #./hyprland.nix
   ];
